requirements.txt
face-recognition
Keras
tensorflow
sklearn
Pillow
opencv-python
